// src/shared/generated/MacroCoordinateRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type MacroCoordinateRecord = { anchor: "top-left" | "top-center" | "top-right" | "center-left" | "center" | "center-right" | "bottom-left" | "bottom-center" | "bottom-right", appliedPageZoom: number, referenceViewportHeightPx: number, referenceViewportWidthPx: number, xPercent: number, xPx: number, xReferencePx: number, viewportHeightPx: number, viewportWidthPx: number, yPercent: number, yPx: number, yReferencePx: number, };